# Conflicts:
#	app/src/main/java/com/qiushi/wechatshop/ui/manage/ManageFragment.kt
#	app/src/main/java/com/qiushi/wechatshop/ui/manage/ManagerAdapter.kt
#	app/src/main/java/com/qiushi/wechatshop/ui/moments/CreateMomentsActivity.kt
#	app/src/main/java/com/qiushi/wechatshop/ui/order/OrderFragment.kt
#	app/src/main/java/com/qiushi/wechatshop/util/Push.java

+++ app/src/main/java/com/qiushi/wechatshop/model/AddGoods.kt
@@ -7,15 +7,13 @@ class AddGoods : Serializable {
     var shop_id: Long = 0
     var name: String = ""
     var cover: String = ""
-    var price: Long = 0
-    var stock: Long = 0
+    var price: Double = 0.00
+    var stock: Int = 0
     var brief: String = ""
     var content: ArrayList<Content>? = null
     var goods_id: Long = 0
     var is_todo: Boolean = false
     var cover_url: String = ""
     var sales_brief: String = ""
-    var left_todo:Int = 0
-
-
+    var left_todo: Int = 0
 }
